requirements-dev.txt
pytest-django==3.0.0
pytest-xdist==1.15.0
pytest-cov==2.3.1
selenium==2.53.6
mock==2.0.0
